feat(knowledge): merge ingestion + retrieval into mode-driven Knowledge component
Merge KnowledgeIngestionComponent and KnowledgeBaseComponent into a single
KnowledgeComponent driven by a TabInput mode selector ("Ingest" / "Retrieve"),
mirroring the PoliciesComponent pattern.

- New: src/lfx/src/lfx/components/files_and_knowledge/knowledge.py holds the
  full merged component. update_build_config and update_outputs hide the
  inputs and the output that do not apply to the current mode.
- Legacy KnowledgeIngestionComponent / KnowledgeBaseComponent become thin
  subclasses pinned to their respective modes with class-level inputs and
  outputs aligned to the legacy shape, so saved flows continue to load.
- Output names ("dataframe_output", "retrieve_data") preserved for back-compat.
- New tests cover mode-switch input/output visibility and subclass pinning;
  existing test_ingestion / test_retrieval suites rerouted to patch the new
  knowledge module path and continue to pass.
